# docs/openapi/resource-center.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new-api Resource Center API",
    "version": "2026-07-17",
    "summary": "Assets, asynchronous image tasks, uploads, and Webhooks.",
    "description": "Public Resource Center API. Async image tasks and uploads use a normal new-api token. Assets use an ak_ resource key. Account-level task Webhooks are configured in the Resource Center console."
  },
...
  "servers": [
    {
      "url": "{base_url}",
      "variables": {
        "base_url": {
          "default": "https://new-api.example.com",
          "description": "Your new-api origin."
        }
      }
    }
  ],
...
  "paths": {
...
    "/v1/image/tasks/{task_id}": {
      "get": {
        "tags": [
          "Async Images"
        ],
        "operationId": "getImageTask",
        "summary": "Get an asynchronous image task",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "Public task ID such as task_xxx.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/v1/image/tasks/query": {
      "post": {
        "tags": [
          "Async Images"
        ],
        "operationId": "queryImageTasks",
        "summary": "Query up to 100 image tasks",
        "description": "Returned tasks preserve request order. Unknown or unauthorized IDs are listed in missing.",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageTaskBatchQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ordered task list with missing IDs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTaskBatchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "new-api token",
        "description": "A normal new-api token authorized for the requested image model."
      },
      "AssetKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_*",
        "description": "A Resource Center key. Legacy keys have assets:read only."
      },
      "WebhookBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wk-...",
        "description": "new-api generates this Key for the account-level task Webhook. The account owner can reveal, copy, or regenerate it in the Resource Center."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "code",
          "message"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "invalid_request_error"
          },
          "code": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
...
      "ImageTaskOperation": {
        "type": "string",
        "enum": [
          "generation",
          "edit"
        ]
      },
      "ImageTaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "in_progress",
          "succeeded",
          "failed"
        ]
      },
...
      "ImageTaskResultImage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "asset_id",
          "url"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "mime_type": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "width": {
            "type": "integer",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "minimum": 0
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "filename": {
            "type": "string"
          },
          "revised_prompt": {
            "type": "string"
          }
        }
      },
      "ImageTaskResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "images"
        ],
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageTaskResultImage"
            }
          },
          "output": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ImageTaskError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ImageTask": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "object",
          "model",
          "operation",
          "status",
          "progress",
          "result",
          "usage",
          "error",
          "created_at",
          "started_at",
          "completed_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^task_"
          },
          "object": {
            "const": "image.task"
          },
          "model": {
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/ImageTaskOperation"
          },
          "status": {
            "$ref": "#/components/schemas/ImageTaskStatus"
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "result": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ImageTaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "type": "object",
            "additionalProperties": true
          },
          "error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ImageTaskError"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_reference_id": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "created_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "started_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "completed_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ImageTaskBatchQueryRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "task_ids"
        ],
        "properties": {
          "task_ids": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ImageTaskBatchResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "object",
          "data"
        ],
        "properties": {
          "object": {
            "const": "list"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageTask"
            }
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credential.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Credential, scope, IP, expiry, or user policy denied access.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource was not found for the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}